+++ src/storage/RecordStore.cpp
@@ -186,7 +186,7 @@ uint32_t nextId() {
     return g_nextId;
 }
 
-bool readLatest(uint16_t offset, uint8_t limit, ReadCallback callback, void* user) {
+bool readLatest(uint16_t offset, uint16_t limit, ReadCallback callback, void* user) {
     if (!callback || !g_ready || offset >= g_count) {
         return false;
     }
